feat(backend): Image provider is a service
The service has a fale implementation within the debug environment.
In the other cases, the base URL is fetched from the environment which, in turn, is fed by the deploy workflow by means of secrets.

New environment variables must be set in order for the application to run (otherwise a runtime error occurs at application startup).
  - IMAGES_BASE_URL
  - STAGE