breaking(backend): product attributes update
The product description has been updated and images correctly associated.
The product has the new attribute `DescTitle` which contains the title of the extended description.
The description contains the `\n` character, that must be replaced by a `CR` in the rendering phase, both on the GUi and on the pdf quote.