feat(backend): supplier logo is rendered on pdf quotation
The logo is placed in tha same place where the supplier data where located. When a logo is specified by the supplier, it is rendered in place of the supplier data. Supplier data are rendered as fallback in case no logo is provided by the supplier.

The logo has a maximum (width, height) equal to (320, 150).